move factory code to its own module
