+++ rag_experiment_accelerator/checkpoint/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa: F401
 
-from rag_experiment_accelerator.checkpoint.checkpoint import (
+from rag_experiment_accelerator.checkpoint.checkpoint_factory import (
     init_checkpoint,
     get_checkpoint,
 )
