+++ rag_experiment_accelerator/checkpoint/tests/test_checkpoint.py
@@ -2,7 +2,7 @@ from unittest.mock import MagicMock
 import pytest
 from unittest.mock import patch
 
-from rag_experiment_accelerator.checkpoint.checkpoint import (
+from rag_experiment_accelerator.checkpoint.checkpoint_factory import (
     get_checkpoint,
     init_checkpoint,
 )
